requirements.txt
jupyter
pandas
matplotlib
jupytext
ipykernel
rtree
geopandas
dateutils
